move the rest of admin twig templates and JS files from project-base to the framework package
