+++ packages/framework/assets/js/admin/components/Product.js
@@ -3,6 +3,7 @@ import Register from '../../common/utils/Register';
 export default class Product {
     static init ($container) {
         Product.initializeSideNavigation($container);
+        Product.initProductVideos($container);
     }
 
     static initializeSideNavigation ($container) {
@@ -21,6 +22,28 @@ export default class Product {
                 const scrollOffsetTop = $title.offset().top - $webContent.offset().top;
                 $('html, body').animate({ scrollTop: scrollOffsetTop }, 'slow');
             });
+        });
+    }
+
+    static initProductVideos ($container) {
+        $container.filterAllNodes('.js-videos-collection').on('click', '.js-remove-row', function () {
+            $(this).parent().parent().remove();
+        });
+
+        $container.filterAllNodes('.js-videos-collection-add-row').on('click', function (event) {
+            const $collection = $(this).closest('.js-form-group').find('.js-videos-collection');
+            let index = $collection.data('index');
+            index++;
+            let prototype = $collection.data('prototype');
+            let item = prototype
+                .replace(/__name__label__/g, index)
+                .replace(/__name__/g, index);
+
+            let $item = $($.parseHTML(item));
+
+            $item.data('index', index);
+            $collection.data('index', index);
+            $collection.append($item);
         });
     }
 }
