+++ packages/framework/src/Resources/views/Admin/Content/Slider/detail.html.twig
@@ -1,12 +1,7 @@
 {% extends '@ShopsysFramework/Admin/Layout/layoutWithPanel.html.twig' %}
 
-{% block title %}
-    - {{ 'Slider page'|trans }}
-{% endblock %}
-
-{% block h1 %}
-    {{ 'Slider page'|trans }}
-{% endblock %}
+{% block title %}- {{ 'Banner'|trans }}{% endblock %}
+{% block h1 %}{{ 'Banner'|trans }}{% endblock %}
 
 {% block main_content %}
