+++ packages/framework/src/Resources/views/Admin/Content/Slider/edit.html.twig
@@ -1,4 +1,4 @@
 {% extends '@ShopsysFramework/Admin/Content/Slider/detail.html.twig' %}
 
-{% block title %}- {{ 'Editing slider page'|trans }} - {{ sliderItem.name }}{% endblock %}
-{% block h1 %}{{ 'Editing slider page'|trans }} - {{ sliderItem.name }}{% endblock %}
+{% block title %}- {{ 'Editing banner'|trans }} - {{ sliderItem.name }}{% endblock %}
+{% block h1 %}{{ 'Editing banner'|trans }} - {{ sliderItem.name }}{% endblock %}
