+++ packages/framework/src/Resources/views/Admin/Content/Slider/new.html.twig
@@ -1,4 +1,4 @@
 {% extends '@ShopsysFramework/Admin/Content/Slider/detail.html.twig' %}
 
-{% block title %}- {{ 'New slider page'|trans }}{% endblock %}
-{% block h1 %}{{ 'New slider page'|trans }}{% endblock %}
+{% block title %}- {{ 'New banner'|trans }}{% endblock %}
+{% block h1 %}{{ 'New banner'|trans }}{% endblock %}
