+++ packages/framework/src/Resources/views/Admin/Content/TopCategory/list.html.twig
@@ -1,7 +1,7 @@
 {% extends '@ShopsysFramework/Admin/Layout/layoutWithPanel.html.twig' %}
 
-{% block title %}- {{ 'Category guidepost on main page'|trans }}{% endblock %}
-{% block h1 %}{{ 'Category guidepost on main page'|trans }}{% endblock %}
+{% block title %}- {{ 'Promoted categories'|trans }}{% endblock %}
+{% block h1 %}{{ 'Promoted categories'|trans }}{% endblock %}
 
 {% block main_content %}
     {{ render(controller('Shopsys\\FrameworkBundle\\Controller\\Admin\\DomainController::domainTabsAction')) }}
