+++ packages/framework/src/Resources/views/Admin/Content/TopProduct/list.html.twig
@@ -1,7 +1,7 @@
 {% extends '@ShopsysFramework/Admin/Layout/layoutWithPanel.html.twig' %}
 
-{% block title %}- {{ 'Main page products'|trans }}{% endblock %}
-{% block h1 %}{{ 'Main page products'|trans }}{% endblock %}
+{% block title %}- {{ 'Promoted products'|trans }}{% endblock %}
+{% block h1 %}{{ 'Promoted products'|trans }}{% endblock %}
 
 {% block main_content %}
     {{ render(controller('Shopsys\\FrameworkBundle\\Controller\\Admin\\DomainController::domainTabsAction')) }}
