+++ packages/framework/src/Resources/views/Admin/Content/TransportAndPayment/list.html.twig
@@ -9,6 +9,7 @@
 {% endblock %}
 
 {% block main_content %}
+    {{ render(controller('Shopsys\\FrameworkBundle\\Controller\\Admin\\DomainController::domainTabsAction')) }}
     {{ render(controller('Shopsys\\FrameworkBundle\\Controller\\Admin\\TransportController::listAction')) }}
     {{ render(controller('Shopsys\\FrameworkBundle\\Controller\\Admin\\PaymentController::listAction')) }}
