+++ packages/framework/src/Resources/views/Admin/Content/Unit/list.html.twig
@@ -1,7 +1,7 @@
 {% extends '@ShopsysFramework/Admin/Layout/layoutWithPanel.html.twig' %}
 
-{% block title %}- {{ 'Units overview'|trans }}{% endblock %}
-{% block h1 %}{{ 'Units overview'|trans }}{% endblock %}
+{% block title %}- {{ 'Measurement units overview'|trans }}{% endblock %}
+{% block h1 %}{{ 'Measurement units overview'|trans }}{% endblock %}
 
 {% block main_content %}
     <div class="wrap-divider wrap-divider--bottom">
