+++ packages/framework/src/Resources/views/Admin/Form/Product/productDetailFormTheme.html.twig
