+++ project-base/app/assets/js/admin/admin.js
@@ -2,12 +2,6 @@ import '../jQuery/registerJquery';
 import registerAdmin from 'framework/admin/registerAdmin';
 import '../loadTranslations';
 
-import './validation';
-import './advert/advert';
-import './../common/validation/customizeFpValidator';
-
 import './grapesjs/initGrapesJs';
-
-import './product/product';
 
 registerAdmin();
