--- project-base/app/assets/js/admin/validation/index.js
@@ -1,1 +1,0 @@
-import './promoCodeValidator';
