+++ project-base/app/config/packages/twig.yaml
@@ -5,6 +5,8 @@ twig:
         - '@ShopsysFramework/Admin/Form/Product/Price/priceListOverviewType.html.twig'
         - '@ShopsysFramework/Admin/Form/Product/Price/pricesByPricingGroupsType.html.twig'
         - '@ShopsysFramework/Admin/Form/Product/Price/productPricesWithVatSelectType.html.twig'
+        - '@ShopsysFramework/Admin/Form/Product/productDetailFormTheme.html.twig'
+        - '@ShopsysFramework/Admin/Form/Product/productVideoValue.html.twig'
         - '@ShopsysFramework/Admin/Form/abstractFileuploadFields.html.twig'
         - '@ShopsysFramework/Admin/Form/abstractMultiplePicker.html.twig'
         - '@ShopsysFramework/Admin/Form/blogCategoriesType.html.twig'
@@ -44,8 +46,6 @@ twig:
         - '@ShopsysFramework/Admin/Form/rolesType.html.twig'
         - '@ShopsysFramework/Admin/Form/storeOpeningHoursFormTheme.html.twig'
         - '@ShopsysFramework/Admin/Form/transportPricesFormTheme.html.twig'
-        - 'Admin/Form/productDetailFormTheme.html.twig'
-        - 'Admin/Form/productVideoValue.html.twig'
     globals:
         globalMultipleFormSetting: '@Shopsys\FrameworkBundle\Component\Form\MultipleFormSetting'
         usersnapApiKey: '%env(USERSNAP_PROJECT_API_KEY)%'
