--- project-base/app/templates/bundles/ShopsysFrameworkBundle/Admin/Content/Slider/detail.html.twig
@@ -1,4 +1,0 @@
-{% extends '@!ShopsysFramework/Admin/Content/Slider/detail.html.twig' %}
-
-{% block title %}- {{ 'Banner'|trans }}{% endblock %}
-{% block h1 %}{{ 'Banner'|trans }}{% endblock %}
