--- project-base/app/templates/bundles/ShopsysFrameworkBundle/Admin/Content/Slider/edit.html.twig
@@ -1,4 +1,0 @@
-{% extends '@!ShopsysFramework/Admin/Content/Slider/edit.html.twig' %}
-
-{% block title %}- {{ 'Editing banner'|trans }} - {{ sliderItem.name }}{% endblock %}
-{% block h1 %}{{ 'Editing banner'|trans }} - {{ sliderItem.name }}{% endblock %}
