--- project-base/app/templates/bundles/ShopsysFrameworkBundle/Admin/Content/Slider/list.html.twig
@@ -1,12 +1,0 @@
-{% extends '@!ShopsysFramework/Admin/Content/Slider/list.html.twig' %}
-
-{% block title %}- {{ 'Banners'|trans }}{% endblock %}
-{% block h1 %}{{ 'Banners'|trans }}{% endblock %}
-
-{% block btn %}
-    <div class="wrap-bar__buttons">
-        <a href="{{ url('admin_slider_new') }}" class="btn btn--primary btn--plus wrap-bar__btn">
-            <i class="btn__icon">+</i>{{ 'Create new banner'|trans }}
-        </a>
-    </div>
-{% endblock %}
