--- project-base/app/templates/bundles/ShopsysFrameworkBundle/Admin/Content/Slider/new.html.twig
@@ -1,4 +1,0 @@
-{% extends '@!ShopsysFramework/Admin/Content/Slider/new.html.twig' %}
-
-{% block title %}- {{ 'New banner'|trans }}{% endblock %}
-{% block h1 %}{{ 'New banner'|trans }}{% endblock %}
