--- project-base/app/templates/bundles/ShopsysFrameworkBundle/Admin/Content/TopCategory/list.html.twig
@@ -1,4 +1,0 @@
-{% extends '@!ShopsysFramework/Admin/Content/TopCategory/list.html.twig' %}
-
-{% block title %}- {{ 'Promoted categories'|trans }}{% endblock %}
-{% block h1 %}{{ 'Promoted categories'|trans }}{% endblock %}
