--- project-base/app/templates/bundles/ShopsysFrameworkBundle/Admin/Content/TopProduct/list.html.twig
@@ -1,4 +1,0 @@
-{% extends '@!ShopsysFramework/Admin/Content/TopProduct/list.html.twig' %}
-
-{% block title %}- {{ 'Promoted products'|trans }}{% endblock %}
-{% block h1 %}{{ 'Promoted products'|trans }}{% endblock %}
