--- project-base/app/templates/bundles/ShopsysFrameworkBundle/Admin/Content/Unit/list.html.twig
@@ -1,3 +1,0 @@
-{% extends '@!ShopsysFramework/Admin/Content/Unit/list.html.twig' %}
-{% block title %}- {{ 'Measurement units overview'|trans }}{% endblock %}
-{% block h1 %}{{ 'Measurement units overview'|trans }}{% endblock %}
